Module Alpha_context.Fitness

include module type of Tezos_protocol_environment.Fitness

The fitness of a block is defined as a list of bytes, compared in a lexicographical order (longer list are greater).

include Tezos_protocol_environment.S.T with type t = Tezos_protocol_environment.MBytes.t list

type t = Tezos_protocol_environment.MBytes.t list

include Tezos_protocol_environment.Compare.S with type t := t

type t

val (=) : t -> t -> bool

val (<>) : t -> t -> bool

val (<) : t -> t -> bool

val (<=) : t -> t -> bool

val (>=) : t -> t -> bool

val (>) : t -> t -> bool

val compare : t -> t -> int

val equal : t -> t -> bool

val max : t -> t -> t

val min : t -> t -> t

val pp : Tezos_protocol_environment.Format.formatter -> t -> unit

val encoding : t Tezos_protocol_environment.Data_encoding.t

val to_bytes : t -> Tezos_protocol_environment.MBytes.t

val of_bytes : Tezos_protocol_environment.MBytes.t -> t option

type fitness = t

val increase : ?⁠gap:int -> context -> context

val current : context -> int64

val to_int64 : fitness -> int64 Tezos_protocol_environment.Error_monad.tzresult