Module Storage.Roll

Storage from this submodule must only be accessed through the module `Roll`.

module Owner : Storage_sigs.Indexed_data_snapshotable_storage with type key = Roll_repr.t and type snapshot = Cycle_repr.t * int and type value = Tezos_protocol_environment.Signature.Public_key.t and type t := Raw_context.t

val clear : Raw_context.t -> Raw_context.t Tezos_protocol_environment.Lwt.t

module Next : Storage_sigs.Single_data_storage with type value = Roll_repr.t and type t := Raw_context.t

The next roll to be allocated.

Rolls linked lists represent both account owned and free rolls. All rolls belongs either to the limbo list or to an owned list.

module Limbo : Storage_sigs.Single_data_storage with type value = Roll_repr.t and type t := Raw_context.t

Head of the linked list of rolls in limbo

module Delegate_roll_list : Storage_sigs.Indexed_data_storage with type key = Tezos_protocol_environment.Signature.Public_key_hash.t and type value = Roll_repr.t and type t := Raw_context.t

Rolls associated to contracts, a linked list per contract

module Successor : Storage_sigs.Indexed_data_storage with type key = Roll_repr.t and type value = Roll_repr.t and type t := Raw_context.t

Use this to iter on a linked list of rolls

module Delegate_change : Storage_sigs.Indexed_data_storage with type key = Tezos_protocol_environment.Signature.Public_key_hash.t and type value = Tez_repr.t and type t := Raw_context.t

The tez of a contract that are not assigned to rolls

module Snapshot_for_cycle : Storage_sigs.Indexed_data_storage with type key = Cycle_repr.t and type value = int and type t := Raw_context.t

Index of the randomly selected roll snapshot of a given cycle.

module Last_for_snapshot : Storage_sigs.Indexed_data_storage with type key = int and type value = Roll_repr.t and type t = Raw_context.t * Cycle_repr.t

Last roll in the snapshoted roll allocation of a given cycle.