Module type Storage_sigs.Non_iterable_indexed_carbonated_data_storage

Variant of Non_iterable_indexed_data_storage with gas accounting.

type t

type context = t

type key

An abstract type for keys

type value

The type of values

val mem : context -> key -> (Raw_context.t * bool) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Tells if a given key is already bound to a storage bucket. Consumes Gas_repr.read_bytes_cost Z.zero.

val get : context -> key -> (Raw_context.t * value) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Retrieve a value from the storage bucket at a given key ; returns Storage_error Missing_key if the key is not set ; returns Storage_error Corrupted_data if the deserialisation fails. Consumes Gas_repr.read_bytes_cost <size of the value>.

val get_option : context -> key -> (Raw_context.t * value option) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Retrieve a value from the storage bucket at a given key ; returns None if the value is not set ; returns Storage_error Corrupted_data if the deserialisation fails. Consumes Gas_repr.read_bytes_cost <size of the value> if present or Gas_repr.read_bytes_cost Z.zero.

val set : context -> key -> value -> (Raw_context.t * int) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Updates the content of a bucket ; returns A Storage_Error Missing_key if the value does not exists. Consumes serialization cost. Consumes Gas_repr.write_bytes_cost <size of the new value>. Returns the difference from the old to the new size.

val init : context -> key -> value -> (Raw_context.t * int) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Allocates a storage bucket at the given key and initializes it ; returns a Storage_error Existing_key if the bucket exists. Consumes serialization cost. Consumes Gas_repr.write_bytes_cost <size of the value>. Returns the size.

val init_set : context -> key -> value -> (Raw_context.t * int) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Allocates a storage bucket at the given key and initializes it with a value ; just updates it if the bucket exists. Consumes serialization cost. Consumes Gas_repr.write_bytes_cost <size of the new value>. Returns the difference from the old (maybe 0) to the new size.

val set_option : context -> key -> value option -> (Raw_context.t * int) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

When the value is Some v, allocates the data and initializes it with v ; just updates it if the bucket exists. When the valus is None, delete the storage bucket when the value ; does nothing if the bucket does not exists. Consumes serialization cost. Consumes the same gas cost as either remove or init_set. Returns the difference from the old (maybe 0) to the new size.

val delete : context -> key -> (Raw_context.t * int) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Delete a storage bucket and its contents ; returns a Storage_error Missing_key if the bucket does not exists. Consumes Gas_repr.write_bytes_cost Z.zero. Returns the freed size.

val remove : context -> key -> (Raw_context.t * int) Tezos_protocol_environment.Error_monad.tzresult Tezos_protocol_environment.Lwt.t

Removes a storage bucket and its contents ; does nothing if the bucket does not exists. Consumes Gas_repr.write_bytes_cost Z.zero. Returns the freed size.